pull out declarations of l1_config and l1_server to ch_frb_l1.hpp
