+++ ch_frb_l1.hpp
@@ -1,6 +1,7 @@
 #ifndef _CH_FRB_L1_HPP
 #define _CH_FRB_L1_HPP
 
+#include <thread>
 #include <string>
 #include <vector>
 #include <cstring>
@@ -9,10 +10,232 @@
 
 #include <yaml-cpp/yaml.h>
 
+#include <rf_pipelines.hpp>
+#include <bonsai.hpp>
+#include <ch_frb_io_internals.hpp>
+#include <l1-rpc.hpp>
+#include <l1-prometheus.hpp>
+
 namespace ch_frb_l1 {
 #if 0
 }  // compiler pacifier
 #endif
+
+
+// -------------------------------------------------------------------------------------------------
+//
+// l1_config: reads and parses config files, does a lot of sanity checking,
+// but does not allocate any "heavyweight" data structures.
+struct l1_config
+{
+    l1_config() { }
+    l1_config(int argc, char **argv);
+
+    // Command-line arguments
+    std::string l1_config_filename;
+    std::string rfi_config_filename;
+    std::string bonsai_config_filename;
+    std::string l1b_config_filename;
+
+    Json::Value rfi_transform_chain_json;
+    bonsai::config_params bonsai_config;
+
+    // Command-line flags
+    // Currently, l1_verbosity can have the following values (I may add more later):
+    //   1: pretty quiet
+    //   2: pretty noisy
+
+    bool tflag = false;
+    bool fflag = false;
+    bool l1b_pipe_io_debug = false;
+    bool memory_pool_debug = false;
+    bool write_chunk_debug = false;
+    bool deliberately_crash = false;
+    bool ignore_end_of_stream = false;
+    int l1_verbosity = 1;
+
+    // nstreams is automatically determined by the number of (ipaddr, port) pairs.
+    // There will be one (network_thread, assembler_thread, rpc_server) triple for each stream.
+    int nbeams = 0;
+    int nfreq = 0;
+    int nstreams = 0;
+    int nt_per_packet = 0;
+    int fpga_counts_per_sample = 384;
+    int nt_align = 0;   // used to align stream of assembled_chunks to RFI/dedispersion block size.
+
+    // The number of frequencies in the downsampled RFI chain.
+    // Must match the number of downsampled frequencies in the RFI chain JSON file.  (probably 1024)
+    int nrfifreq = 0;
+
+    // If slow_kernels=false (the default), the L1 server will use fast assembly language kernels
+    // for its packet processing.  If slow_kernels=true, then it will use reference kernels which
+    // are much slower.
+    //
+    // Note 1: the slow kernels are too slow for non-subscale use!  If slow kernels are used on
+    // the "production-scale" L1 server with (nbeams, nupfreq) = (16, 16), it may crash.
+    //
+    // Note 2: the fast kernels can only be used if certain conditions are met.  As of this writing,
+    // the conditions are: (a) nupfreq must be even, and (b) nt_per_packet must be 16.  In particular,
+    // for a subscale run with nupfreq=1, the fast kernels can't be used.
+    //
+    // Conditions (a) and (b) could be generalized by writing a little more code, if this would be useful
+    // then let me know!
+    bool slow_kernels = false;
+
+    // Both vectors have length nstreams.
+    std::vector<std::string> ipaddr;
+    std::vector<int> port;
+
+    // One L1-RPC per stream
+    std::vector<std::string> rpc_address;
+    // One L1-prometheus per stream
+    std::vector<std::string> prometheus_address;
+    // Optional chlog logging server address
+    std::string logger_address;
+
+    // Optional URL to get frame0-ctime
+    std::string frame0_url;
+    // Timeout (in ms) for retrieving frame0-ctime
+    int frame0_timeout;
+
+    // Size of RFI mask measurement ringbuffer (15 seconds required for prometheus; more could be useful for other monitoring tools)
+    int rfi_mask_meas_history;
+
+    // A vector of length nbeams, containing the beam_ids that will be processed on this L1 server.
+    // It is currently assumed that these are known in advance and never change!
+    // If unspecified, 'beam_ids' defaults to { 0, ..., nbeams-1 }.
+    std::vector<int> beam_ids;
+
+    // Buffer sizes, as specified in config file.
+    int unassembled_ringbuf_nsamples = 4096;
+    int assembled_ringbuf_nsamples = 8192;
+    std::vector<int> telescoping_ringbuf_nsamples;
+    double write_staging_area_gb = 0.0;
+    
+    // "Derived" unassembled ringbuf parameters.
+    int unassembled_ringbuf_capacity = 0;
+    int unassembled_nbytes_per_list = 0;
+    int unassembled_npackets_per_list = 0;
+
+    // "Derived" assembled and telescoping ringbuf parameters.
+    int assembled_ringbuf_nchunks = 0;
+    std::vector<int> telescoping_ringbuf_nchunks;
+
+    // Parameters of the memory_slab_pool(s)
+    int nbytes_per_memory_slab = 0;     // size (in bytes) of memory slab used to store assembled_chunk
+    int total_memory_slabs = 0;         // total for node
+
+    // List of output devices (e.g. '/ssd', '/nfs')
+    std::vector<std::string> output_device_names;
+
+    // If 'intensity_prescale' is specified, then all intensity values will be multiplied by its value.
+    // This is a workaround for 16-bit overflow issues in bonsai.  When data is saved to disk, the
+    // prescaling will not be applied.
+    float intensity_prescale = 1.0;
+
+    // L1b linkage.  Note: assumed L1b command line is:
+    //   <l1b_executable_filename> <l1b_config_filename> <beam_id>
+
+    std::string l1b_executable_filename;
+    bool l1b_search_path = false;     // will $PATH be searched for executable?
+    int l1b_buffer_nsamples = 0;      // determines buffer size between L1a and L1b (0 = system default)
+    double l1b_pipe_timeout = 0.0;    // timeout in seconds between L1a and L1
+    bool l1b_pipe_blocking = false;   // setting this to true is equivalent to a very large l1b_pipe_timeout
+
+    // "Derived" parameter: L1b pipe capacity (derived from l1b_buffer_nsamples)
+    int l1b_pipe_capacity = 0;
+
+    // Forces RFI removal and dedispersion to run in separate threads (shouldn't
+    // need to specify this explicitly, except for debugging).
+    bool force_asynchronous_dedispersion = false;
+
+    // Occasionally useful for debugging: If track_global_trigger_max is true, then when 
+    // the L1 server exits, it will print the (DM, arrival time) of the most significant FRB.
+    bool track_global_trigger_max = false;
+
+    // Also intended for debugging.  If the optional parameter 'stream_acqname' is
+    // specified, then the L1 server will auto-stream all chunks to disk.  Warning:
+    // it's very easy to use a lot of disk space this way!
+    
+    std::string stream_devname;            // specified in config file, options are "ssd" or "nfs", defaults to "ssd"
+    std::string stream_acqname;            // specified in config file, defaults to "", which results in no streaming acquisition
+    std::vector<int> stream_beam_ids;      // specified in config file, defaults to all beam ID's on node
+    std::string stream_filename_pattern;   // derived from 'stream_devname' and 'stream_acqname'
+
+    void _have_warnings() const;
+};
+
+
+// -------------------------------------------------------------------------------------------------
+//
+// This master data structure defines an L1 server instance.
+
+
+struct l1_server {        
+    using corelist_t = std::vector<int>;
+
+    const l1_config config;
+    
+    int ncpus = 0;
+    std::vector<corelist_t> cores_on_cpu;   // length-ncpus, defines mapping of cores to cpus
+    
+    // Core-pinning scheme
+    corelist_t output_thread_cores;             // assumed same corelist for all I/O threads
+    corelist_t network_thread_cores;            // assumed same for all network threads
+    std::vector<corelist_t> dedispersion_cores;      // length config.nbeams, used for (RFI, dedisp, L1B).
+    std::vector<corelist_t> assembler_thread_cores;  // length nstreams
+    bool asynchronous_dedispersion = false;     // run RFI and dedispersion in separate threads?
+    double sleep_hack = 0.0;                    // a temporary kludge that will go away soon
+
+    std::string command_line;
+
+    // "Heavyweight" data structures.
+    std::vector<std::shared_ptr<bonsai::trigger_pipe>> l1b_subprocesses;   // can be vector of empty pointers, if L1B is not being run.
+    std::vector<std::shared_ptr<ch_frb_io::output_device>> output_devices;
+    std::vector<std::shared_ptr<ch_frb_io::memory_slab_pool>> memory_slab_pools;
+    std::vector<std::shared_ptr<ch_frb_io::intensity_network_stream>> input_streams;
+
+    std::vector<std::shared_ptr<mask_stats_map> > mask_stats_maps;
+    std::vector<std::shared_ptr<L1RpcServer>> rpc_servers;
+    std::vector<std::shared_ptr<L1PrometheusServer>> prometheus_servers;
+    std::vector<std::thread> rpc_threads;
+    std::vector<std::thread> dedispersion_threads;
+    std::mutex bonsai_dedisp_mutex;
+    std::vector<std::shared_ptr<const bonsai::dedisperser> > bonsai_dedispersers;
+
+    std::vector<std::shared_ptr<const rf_pipelines::pipeline_object> > latency_monitors_pre;
+    std::vector<std::shared_ptr<const rf_pipelines::pipeline_object> > latency_monitors_post;
+
+    // The constructor reads the configuration files, does a lot of sanity checks,
+    // but does not initialize any "heavyweight" data structures.
+    l1_server(int argc, char **argv);
+
+    // These methods incrementally construct the "heavyweight" data structures.
+    void start_logging();
+    void spawn_l1b_subprocesses();
+    void make_output_devices();
+    void make_memory_slab_pools();
+    void make_input_streams();
+    void make_mask_stats();
+    void make_rpc_servers();
+    void make_prometheus_servers();
+    void spawn_dedispersion_threads();
+
+    // These methods wait for the server to exit, and print some summary info.
+    void join_all_threads();
+    void print_statistics();
+
+    // Helper methods called by constructor.
+    void _init_subscale();
+    void _init_20cores_16beams();
+    void _init_20cores_8beams();
+
+    // Called-back by dedispersion thread
+    void set_bonsai(int ibeam, std::shared_ptr<const bonsai::dedisperser>,
+                    std::shared_ptr<const rf_pipelines::pipeline_object> latency_pre,
+                    std::shared_ptr<const rf_pipelines::pipeline_object> latency_post);
+};
+
 
 
 // -------------------------------------------------------------------------------------------------
